+++ src/devices/src/lib.rs
@@ -14,7 +14,7 @@ pub mod pseudo;
 pub mod virtio;
 
 pub use self::bus::{Bus, BusDevice, Error as BusError};
-use crate::virtio::QueueError;
+use crate::virtio::{QueueError, VsockError};
 use logger::{error, IncMetric, METRICS};
 
 // Function used for reporting error in terms of logging
@@ -43,4 +43,6 @@ pub enum Error {
     MalformedDescriptor,
     /// Error during queue processing.
     QueueError(QueueError),
+    /// Vsock device error.
+    VsockError(VsockError),
 }